% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\docType{data}
\name{complete}
\alias{complete}
\title{Complete Separation Example of Agresti}
\format{
Variables are y the response vector (zero-or-one-valued) and x the (quantitative) predictor vector.

The formula y ~ x describes the model.
}
\source{
Agresti (2013), Section 6.5.1.
}
\usage{
complete
}
\description{
Simple logistic regression data for which 
the maximum likelihood estimate does not exist in the conventional sense.
}
\references{
Agresti, A. (2013) Categorical Data Analysis, third edition. Hoboken, NJ: John Wiley \& Sons
}
