% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.glmdr}
\alias{predict.glmdr}
\title{Exponential Family Generalized Linear Models Done Right}
\usage{
predict(object, newdata = new_df, cirt="AICc")
}
\arguments{
\item{object}{a fitted object of class \code{glmdr}.}

\item{newdata}{An optional data frame having predictors. If omitted, the fitted predictors are used.}

\item{crit}{Information Criteria, can be one of "AIC", "BIC", or "AICc."}

\item{alpha}{the confidence level. The default is set to \code{alpha = 0.05}.}

\item{interval}{eight different methods to obtain a confidence interval. Default is a Wilson interval. See also \link[binom]{binom.confint}}
}
\value{
A dataframe that includes (1 - alpha) times 100 percent confidence intervals for mean-value parameters for new data points.
}
\description{
Provide the prediction as a one-sided confidence intervals given new data.
Not appropriate when the MLE exists in the original model.
}
\examples{
\dontrun{
mod <- glmdr(y~x, family="binomial")
predict(mod, newdata = new_df, cirt="AICc")
}
}
