% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi.R
\docType{data}
\name{quasi}
\alias{quasi}
\title{Quasi-Complete Separation Example of Agresti}
\format{
Variables are \code{y} the response vector (zero-or-one-valued)
and \code{x} the (quantitative) predictor vector.

The formula \code{y ~ x} describes the model.
}
\source{
Agresti (2013), Section 6.5.1.
}
\usage{
quasi
}
\description{
Simple logistic regression data for which the maximum likelihood estimate
does not exist in the conventional sense.
}
\references{
Agresti, A. (2013)
\emph{Categorical Data Analysis}, third edition.
Hoboken, NJ: John Wiley \& Sons
}
