% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmdr.R
\name{glmdr}
\alias{glmdr}
\title{Exponential Family Generalized Linear Models Done Right}
\usage{
glmdr(formula, family = c("binomial", "poisson"), data, subset, na.action, offset, contrasts = NULL, 
                  tolerance = c(1e-8,1e-8))
}
\arguments{
\item{formula}{an object of class \code{"\link{formula}"} (or one that 
can be coerced to that class): a symbolic description of the
model to be fitted. See \code{\link[stats]{glm}} and \code{\link[stats]{formula}} 
for description of the R formula mini-language.}

\item{family}{character string specifying the family, must be one of
\code{"binomial"} (default) or \code{"poisson"}.  May be abbreviated.}

\item{data}{an optional data frame, list or environment (or object 
coercible by \code{\link{as.data.frame}} to a data frame) containing
the variables in the model.  If not found in \code{data}, the
variables are taken from \code{environment(formula)},
typically the environment from which \code{glmdr} is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen
when the data contain \code{NA}s.  The default is set by
the \code{na.action} setting of \code{\link{options}}, and is
\code{\link{na.fail}} if that is unset.  The \sQuote{factory-fresh}
default is \code{\link{na.omit}}.  Another possible value is
\code{NULL}, no action.  Value \code{\link{na.exclude}} can be useful.}

\item{offset}{this can be used to specify an \emph{a priori} known 
component to be included in the linear predictor during fitting.
This should be \code{NULL} or a numeric vector of length equal to
the number of cases.  One or more \code{\link{offset}} terms can be
included in the formula instead or as well, and if more than one is
specified their sum is used.  See \code{\link{model.offset}}.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{model.matrix.default}.}

\item{tolerance}{Optimization tolerances for: 1. determining null vectors of Fisher information; 
2. determining the linearity (the responses which are not constrained to be their observed value)}
}
\value{
\code{glmdr} returns an object of class inheriting from \code{"glmdr"}.
An object of class \code{"glmdr"} is a list containing some or all of the
following components:
  \item{om}{the result of invoking \code{\link{glm}} with the arguments
 to this function.  May or may not be valid.  See \code{\link[stats]{glm}}
 for the components of this object.}
 \item{linearity}{a logical vector that says which components of the response
 are free (not conditioned to be equal to their observed values) in the
 LCM.  Not present when the MLE exists in the OM.}
 \item{lcm}{the result of invoking \code{\link{glm}} with the arguments
  to this function and argument \code{subset} set to provide the
  conditioning of the LCM.  See \code{\link[stats]{glm}}
  for the components of this object.
  Not present when the MLE exists in the OM or when all components
  of the component \code{linearity} are \code{FALSE}, in which case
  the LCM is completely degenerate, conditioning all components of
  the response vector equal to their observed values, and has no
  identifiable parameters to estimate.}
\item{nulls}{a matrix, the columns of which span the null space of the
  Fisher information matrix for the OM evaluated at the MLE for the LCM
  (a limit in which some parameters have gone to infinity).
  This determines the LCM, as explained in Eck and Geyer (submitted),
  but does not need to be interpreted by ordinary users.
  Not present when \code{lcm} not present.}
\item{modmat}{The model matrix corresponding to the \code{om} model fit.}
\item{family}{The exponential family supplied by the user in the 
  \code{family} argument.}
\item{y}{The response vector.}
}
\description{
Fit discrete exponential family generalized linear models (logistic 
regression or Poisson regression with log link), correctly, even when 
the maximum likelihood estimate does not exist in the conventional sense.
Log-linear models for categorical data analysis are also fit correctly
by considering them as Poisson regression.
}
\examples{
# example of section 2.4 in Geyer (2009)
data(sports)
out <- glmdr(cbind(wins, losses) ~ 0 + ., family = "binomial", data = sports)
inference(out)
}
\references{
Geyer, C. J. (2009)
Likelihood inference in exponential families and directions of recession.
\emph{Electronic Journal of Statistics}, \bold{3}, 259--289.

Eck, D. J. and Geyer, C. J. (submitted)
Computationally efficient likelihood inference
in exponential families when the maximum likelihood estimator
does not exist. \url{https://arxiv.org/abs/1803.11240}
}
\seealso{
\code{\link{summary.glmdr}}, etc.
}
