% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Exponential Family Generalized Linear Models Done Right}
\usage{
inference(object, alpha = 0.05, eps = 1e-10)
}
\arguments{
\item{object}{a fitted object of class \code{glmdr}.}

\item{alpha}{the confidence level. The default is set to \code{alpha = 0.05}.}

\item{eps}{A user-specified error tolerance in nloptr. The default is set to \code{eps = 1e-10}.}
}
\value{
A dataframe that includes (1 - alpha) times 100 percent confidence intervals for mean-value parameters.
}
\description{
Computes one-sided confidence intervals corresponding to which components of 
the response are not free (conditioned to be equal to their observed values) 
in the LCM.  Not appropriate when the MLE exists in the original model.
}
\details{
In an exponential family generalized linear model (GLM) the maximum likelihood estimate need not exist.
This function detects this situation and does the right thing in that case. 
For the binomial and Poisson models fit by this function the MLE always exists in 
the Barndorff-Nielsen completion of the original model (OM), and
is always the MLE in the limiting conditional model (LCM), which
conditions the OM on some components of the response vector being
equal to their observed values.

An LCM can be thought of in two ways.  It is obtained by conditioning
the OM as just described.  It is also obtained by taking limits in
the OM as parameters go to infinity.  See Geyer (2009) for further
description.

The function \code{glmdr} detects whether the MLE is in the OM or in 
an LCM, determines which LCM (which components of the response vector
are conditioned), and fits the MLE in the LCM.
This function computes one-sided confidence intervals corresponding to 
which components of the response are not free (conditioned to be equal 
to their observed values) in the LCM.  This function is not appropriate 
when the MLE exists in the OM.
}
\examples{
data(sports)
out <- glmdr(cbind(wins, losses) ~ 0 + ., family = "binomial", data = sports)
inference(out)
}
\references{
Geyer, C. J. (2009)
Likelihood inference in exponential families and directions of recession.
\emph{Electronic Journal of Statistics}, \bold{3}, 259--289.

Eck, D. J. and Geyer, C. J. (2021)
Computationally efficient likelihood inference
in exponential families when the maximum likelihood estimator
does not exist. \emph{Electronic Journal of Statistics}, \bold{15(1)}, 2105--2156.
}
